requirements.txt
Flask==3.1.0
Pillow==11.0.0
numpy==2.0.2
tensorflow==2.18.0